% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyForecast.R
\name{dailyForecast}
\alias{dailyForecast}
\title{dailyForecast}
\usage{
dailyForecast(lat, long)
}
\arguments{
\item{lat}{Latitude of the site of interest}

\item{long}{Longitude of the site of interest}
}
\value{
A data frame object that contains a forecasted rainfall (mm),
 mean temperature (celsius), max temperature (celsius), min temperature (celsius),
 dewpoint (celsius), relative humidity (percent), and estimates of evapotranspiration using
 the McGuinness-Bordne Formulation.
}
\description{
This function will download daily forecast data from weather.gov for a given lat long.
}
\examples{
hourlyForecast(lat = 38.966754, long = -76.569252)

}
