% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourlyForecast.R
\name{hourlyForecast}
\alias{hourlyForecast}
\title{hourlyForecast}
\usage{
hourlyForecast(lat, long)
}
\arguments{
\item{lat}{Latitude of the site of interest}

\item{long}{Longitude of the site of interest}
}
\value{
A data frame object that contains a forecasted rainfall (mm)
 temperature (celsius), dewpoint (celsius), and relative humidity (percent).
}
\description{
This function will download hourly forecast data from weather.gov for a given lat long.
}
\examples{
hourlyForecast(lat = 38.966754, long = -76.569252)

}
